% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{sig_story}
\alias{sig_story}
\title{Sigstory}
\usage{
sig_story(sigstory, outdir)
}
\arguments{
\item{sigstory}{a 'signature_analysis' s3 class object containing
summarised results from a mutational signature analysis.}
}
\value{
Run for its side effects. Invisibly returns NULL.
}
\description{
Sigstory
}
